Getting back the plot command somehow in running state.
Will have to fix the generation of the filename for the plots though.
