+++ f1d.go
@@ -1,6 +1,12 @@
 package galo
 
-import "math"
+import (
+	"image/color"
+	"math"
+
+	"go-hep.org/x/hep/hplot"
+	"gonum.org/v1/plot/plotter"
+)
 
 func Moyal(x, cst, mu, sigma float64) float64 {
 	v := (x - mu) / sigma
@@ -87,4 +93,32 @@ func Levy(x, mu, sigma float64) float64 {
 func Gaus(x, cst, mu, sigma float64) float64 {
 	v := (x - mu) / sigma
 	return cst * math.Exp(-0.5*v*v)
+}
+
+func landauFunc(mu, sigma float64, r, g, b uint8) *plotter.Function {
+	f := plotter.NewFunction(func(x float64) float64 {
+		return Landau(x, mu, sigma) / sigma
+	})
+	f.Color = color.RGBA{R: r, B: b, G: g, A: 255}
+	f.Samples = 1000
+	return f
+}
+
+func levyFunc(mu, sigma float64, r, g, b uint8) *plotter.Function {
+	f := plotter.NewFunction(func(x float64) float64 {
+		return Levy(x, mu, sigma) / sigma
+	})
+	f.Color = color.RGBA{R: r, B: b, G: g, A: 255}
+	f.Samples = 1000
+	return f
+}
+
+func SaveFunction(outputFileName string) {
+	p := hplot.New()
+	p.Add(landauFunc(0.25, 0.2, 0, 0, 255))
+	p.X.Max = 10
+	p.X.Min = -10
+	p.Y.Max = 10
+	p.Y.Min = -10
+	SavePlot(p, outputFileName, "landau")
 }
